Add files via upload
I'm excited to share my very first web scraping project as part of my journey towards becoming an IBM Data Analyst. In this project, I've taken on the task of web scraping data comparing Tesla and GameStop stocks using the powerful tool Beautiful Soup.

This project is not just an assignment; it's a hands-on experience that's allowing me to put my newly acquired skills to the test. I'm learning how to navigate the intricacies of web scraping and working with real-world data – a crucial skill in the data analysis field.

It's an opportunity to blend my passion for data with the tools and techniques I'm learning through the IBM Data Analysts certification. Extracting data from websites and transforming it into a usable format is a bit like being a modern-day digital detective, and I'm excited to explore this world for the first time.

This project is the first step on my journey towards becoming a skilled data analyst, and it's a testament to the practical skills I'm gaining through this certification program. I can't wait to see the insights and stories this data will reveal.