Add files via upload
For my second assignment in the 'Databases and SQL for Data Science with Python' course by IBM, I delved into the world of SQL with Python using SQLite. This was a new and exciting challenge for me, and I want to express my gratitude to IBM for providing such meaningful learning experiences.

In this project, I embarked on a journey to answer thought-provoking questions posed by IBM using SQL queries within a Python environment. The integration of SQL and Python was a revelation for me, as it opened up a world of possibilities for data manipulation and analysis.

This assignment not only allowed me to put my SQL skills to the test but also showcased the real-world applicability of these tools. I was able to dig into data, extract valuable insights, and provide meaningful answers to IBM's questions.

This course has been instrumental in expanding my horizons as a data science enthusiast, and I'm excited to continue this journey, equipped with the knowledge gained from this assignment. I can't wait to see where SQL and Python will lead me next!